[cmake]: includes after libraries are found (LibRaw, FFTW, LCMS2) [build]: fix few compiler warning [Autolevels]: better computing of histogram min/max also using RGB channels

+++ src/HdrCreation/debevec.cpp
@@ -94,7 +94,7 @@ void DebevecOperator::computeFusion(ResponseCurve& response, const WeightFunctio
     float minAllowedValue = weight.minTrustedValue();
 
 #pragma omp parallel for
-    for (int idx = 0; idx < tempFrame.size(); ++idx)
+    for (size_t idx = 0; idx < tempFrame.size(); ++idx)
     {
         // data...
         ColorData redData;
